Remove datepicker arrows on mobile
