+++ assets/js/dashboard/realtime.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import Datepicker from './datepicker'
-import DatepickerArrows from './datepicker-arrows'
 import SiteSwitcher from './site-switcher'
 import Filters from './filters'
 import VisitorGraph from './stats/visitor-graph'
@@ -37,10 +36,7 @@ class Realtime extends React.Component {
               <SiteSwitcher site={this.props.site} loggedIn={this.props.loggedIn} currentUserRole={this.props.currentUserRole} />
               <Filters className="flex" site={this.props.site} query={this.props.query} history={this.props.history} />
             </div>
-            <div className="flex ml-auto pl-2">
-              <DatepickerArrows site={this.props.site} query={this.props.query} />
-              <Datepicker className="w-28 sm:w-36 md:w-44" site={this.props.site} query={this.props.query} />
-            </div>
+            <Datepicker site={this.props.site} query={this.props.query} />
           </div>
         </div>
         <VisitorGraph site={this.props.site} query={this.props.query} timer={this.props.timer} />
